# static/api-spec.yaml
openapi: "3.0.3"
info:
  title: Hive API
  description: |
    Multi-provider LLM proxy with telemetry-driven routing.
    Proxies OpenAI-compatible `/v1/chat/completions` through scored provider pools.
    Includes a management API and real-time WebSocket telemetry stream.
  version: "0.1.0"
servers:
  - url: http://127.0.0.1:8153
    description: Default local server
paths:
...
  /v1/chat/completions:
    post:
      summary: Proxy chat completion
      description: |
        Forwards an OpenAI-compatible chat completion request through the scored
        provider pool. The backend selects the best available (provider, model)
        node using real-time telemetry, auto-routing around CircuitBreakers and
        unsupported features. The proxy always routes to each provider's
        default model — the client's model field is unused.

        Response is a Server-Sent Events stream (text/event-stream).
      operationId: proxyChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                model:
                  type: string
                  description: Unused by proxy — routing always uses provider default model
                messages:
                  type: array
                  items:
                    $ref: "#/components/schemas/Message"
                tools:
                  type: array
                  description: Function/tool definitions (triggers feature routing)
                max_tokens:
                  type: integer
                temperature:
                  type: number
                stream:
                  type: boolean
                  default: false
      responses:
        "200":
          description: SSE stream from routed provider
          headers:
            Content-Type:
              schema:
                type: string
                example: text/event-stream
        "503":
          description: No configured providers or all providers failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  schemas:
    Message:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
        tool_calls:
          type: array
        tool_call_id:
          type: string
...